refactor(lyrics): split claude_merge tests into sibling file for 1000-line cap
D.1's added tests pushed claude_merge.rs to 1145 lines, over the
airuleset 1000-line cap. CI File Size Check failed. Move the entire
`mod tests` block into a sibling `claude_merge_tests.rs` (mirroring
the routes_tests_spotify.rs split pattern from PR #70).

After split:
- claude_merge.rs: 394 lines (production code only)
- claude_merge_tests.rs: 758 lines (full test suite)
